backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
qrcode[pil]==7.4.2
python-barcode==0.15.1
reportlab==4.1.0
redis==5.0.4
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
boto3==1.34.84
